refactor(mga/pr5): extract Review components + move db.flush to repo layer
- Extract ConfidenceBadge, ReviewCard, ReviewScreenshot, ReviewSkeletonCard
  into src/components/review/ (one component per file, fast-refresh clean)
- Move confidenceBorderClass to confidenceUtils.ts (non-component export
  separation for react-refresh compliance)
- Move db.flush() in classifier_service.py to new write_classifier_suggestions()
  in lineup_repo.py (ORM operations belong in the repo layer)
- Improve E2E tests: add confidence filter interaction, accept-button
  presence check, and resilient skip guards for backend-unavailable CI
